all features done. almost finished

+++ src/routers/tableRouter.ts
@@ -19,8 +19,11 @@ tableRouter.get("/status", isAuthMiddleware, async (req, res) => {
 
 tableRouter.patch("/reserve_table/:id", isAuthMiddleware, async (req: Request<{id: string}>, res) => {
     try {
+        if (!req.user || req.user.role !== "cliente") {
+	        res.status(403).json({ error: "No tienes permisos para realizar esta acción." });
+            return;
+        }
         const id = Number.parseInt(req.params.id);
-
         const cliente = req.user as ClienteFromToken;
         const mesa = await table_service.reserveTable(id, cliente.id);
         res.status(200).json({ data: mesa.id, cliente: cliente.id });
@@ -30,8 +33,12 @@ tableRouter.patch("/reserve_table/:id", isAuthMiddleware, async (req: Request<{i
     }
 });
 
-tableRouter.patch("/free_table/:id", async (req, res) => {
+tableRouter.patch("/free_table/:id", isAuthMiddleware, async (req: Request<{id: string}>, res) => {
     try {
+        if (!req.user || req.user.role !== "admin") {
+	        res.status(403).json({ error: "No tienes permisos para realizar esta acción." });
+            return;
+        }
         const id = Number.parseInt(req.params.id);
         const mesa = await table_service.freeTable(id);
         res.status(200).json({ data: mesa.id, "message": "Mesa liberada" });
@@ -41,8 +48,12 @@ tableRouter.patch("/free_table/:id", async (req, res) => {
     }
 });
 
-tableRouter.post("/add_table", async (req, res) => {
+tableRouter.post("/add_table", isAuthMiddleware, async (req, res) => {
     try {
+        if (!req.user || req.user.role !== "admin") {
+	        res.status(403).json({ error: "No tienes permisos para realizar esta acción." });
+            return;
+        }
         const mesa = await table_service.createTable();
         res.status(201).json({ data: mesa.id, "message": "Mesa creada" });
     } catch (error) {
@@ -51,8 +62,12 @@ tableRouter.post("/add_table", async (req, res) => {
     }
 });
 
-tableRouter.post("/remove_table/:id", async (req, res) => {
+tableRouter.post("/remove_table/:id", isAuthMiddleware, async (req: Request<{id: string}>, res) => {
     try {
+        if (!req.user || req.user.role !== "admin") {
+	        res.status(403).json({ error: "No tienes permisos para realizar esta acción." });
+            return;
+        }
         const id = Number.parseInt(req.params.id);
         const mesa = await table_service.deleteTable(id);
         res.status(201).json({ data: mesa.id, "message": "Mesa eliminada" });
